+++ usr.sbin/bhyve/amd64/ioapic.c
