+++ usr.sbin/bhyve/amd64/ioapic.h
