+++ usr.sbin/bhyve/amd64/pci_irq.c
