+++ usr.sbin/bhyve/amd64/pci_irq.h
