+++ usr.sbin/bhyve/amd64/pci_lpc.c
